💄 Integrate Tailwind CSS and add filter select component
The commit incorporates Tailwind CSS into the React project file structure and adds dependencies to support it. Further, a new filter type select component has been added to the country search container. It also includes the setting up of a new hook useCountryFilter to be defined and used later in the application.

Signed-off-by: Gaba Bartnicka <bartnicka.g@gmail.com>